Code for the remote controller end of the bike project
Messy, repetitive, code. It does seem to work when tested with the Atmega2560 though. Waiting on the Nanos to be delivered. 
Minimal program storage and memory consumption. Gives me hope that this will be a simple task. Plenty of room for improvements/new features.

SPECIFICATIONS FOR THE NANO V3.0:
Sketch uses 2892 bytes (9%) of program storage space. Maximum is 30720 bytes.
Global variables use 138 bytes (6%) of dynamic memory, leaving 1910 bytes for local variables. Maximum is 2048 bytes.